// front-office/src/app/openapi/model/userSearchAdminDTO.ts
/**
 * FinalStake
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserSearchAdminDTO { 
    email?: string;
    enabled?: boolean;
    id?: string;
    locked?: boolean;
    username?: string;
}